refactor move simulation evaluator into simulation namespace
